+++ src/pages/login/Redirect.tsx
@@ -3,6 +3,7 @@ import { useNavigate, useLocation } from "react-router-dom";
 import Loading from "../../components/loading/Loading";
 import styled from "@emotion/styled";
 import { googleRedirect } from "../../services/userService";
+import { handleValidDomain } from "../../utils/JWT";
 
 const Container = styled.div`
   display: flex;
@@ -22,10 +23,14 @@ const Redirect = () => {
     const handleRedirect = async () => {
       try {
         const jwtToken = await googleRedirect();
-        console.log(jwtToken);
-        if (jwtToken) {
+        const isValidDomain = handleValidDomain();
+        if (jwtToken && isValidDomain) {
           navigate("/");
+          return;
         }
+        //유효하지 않은 도메인
+        alert("숙명 Gmail 계정으로만 로그인 가능합니다.");
+        navigate("/login");
       } catch (error) {
         console.error("로그인 실패:", error);
       }
